+++ src/main/java/controller/GameController.java
@@ -1,9 +1,13 @@
 package controller;
 
-import domain.*;
+import domain.ExceptionHandler;
 import domain.deck.TotalDeckGenerator;
+import domain.game.Command;
 import domain.game.Game;
 import domain.game.State;
+import domain.money.Money;
+import domain.money.MoneyManager;
+import domain.money.Profit;
 import domain.user.Player;
 import domain.user.User;
 import domain.user.Users;
