+++ src/main/java/domain/money/Money.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.money;
 
 import domain.game.Result;
