+++ src/test/java/domain/CommandTest.java
@@ -1,5 +1,6 @@
 package domain;
 
+import domain.game.Command;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
