+++ src/test/java/domain/MoneyManagerTest.java
@@ -5,6 +5,9 @@ import domain.card.Number;
 import domain.deck.TotalDeck;
 import domain.game.Game;
 import domain.game.Result;
+import domain.money.Money;
+import domain.money.MoneyManager;
+import domain.money.Profit;
 import domain.user.Player;
 import domain.user.Users;
 import org.junit.jupiter.api.DisplayName;
