+++ src/test/java/domain/MoneyTest.java
@@ -1,5 +1,6 @@
 package domain;
 
+import domain.money.Money;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
